apps/api/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.5

# Redis
redis==5.0.4
celery==5.4.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Scraping
playwright==1.43.0
beautifulsoup4==4.12.3
lxml==5.2.1
httpx==0.27.0

# AI/ML
torch==2.2.2
transformers==4.40.1
sentence-transformers==2.7.0
Pillow==10.3.0
numpy==1.26.4
scikit-learn==1.4.2
hdbscan==0.8.33

# Utilities
tenacity==8.2.3
structlog==24.1.0
orjson==3.10.1

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
